Introduce the announcer routine
The announcer routine will announce whenever an event is starting in all
rooms it's currently joined to.

We do this by checking when the next event is, and if it's in less than
10 minutes, we announce it. If not, we sleep for a bit and try again.

Since there is no easy way right now to distinguish between a DM and a
"room room", we check the member count for now as a crude heuristic.