Fix usage of timer
Looking at the godoc it's not safe to call Reset on a non-stopped timer,
and it's not safe to do it from a different goroutine.

This changes the announcer timer into a channel, that in our Announce
function we select on. This in turn should let us safely stop the timer
and drain the channel.

The logic for sending the event itself is now moved out of the select,
so that when we receive a reset or the timer expired the same logic
triggers. This should guarantees that in that block we have an expired
timer, so we can safely call Reset at any point.