Improve on the logging behaviour
By default zap can only output to one of stderr or stdout. Instead, we
configure a new Core that outputs Error or worse to stderr, and anything
below it to stdout.

This also makes the logging behaviour somewhat configurable, with the
option to chose between console output (which logs the message first,
and then the key/value pairs as JSON) or the JSON output. It's now also
possible to toggle the timestamp on/off which is handy if you're running
the bot under something like systemd which already includes timestamps
of its own.